pyproject.toml
[project]
name = "tp3-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.4.3",
    "litellm>=1.80.10",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]
